Move resolveReferences to helpers
