+++ packages/alfa-dom/src/common/index.ts
@@ -1,1 +1,2 @@
+export * from "./helpers";
 export * from "./predicate";
