+++ packages/alfa-dom/src/tables/helpers.ts
@@ -1,18 +1,15 @@
-import { Equatable } from "@siteimprove/alfa-equatable";
 import { clamp } from "@siteimprove/alfa-math";
 import { parseNonNegativeInteger } from "@siteimprove/alfa-parser";
-import { Predicate } from "@siteimprove/alfa-predicate";
 import { Ok } from "@siteimprove/alfa-result";
 
-import { Element, Node, Attribute } from "..";
+import { Element, Attribute } from "..";
 
 const { parseAttribute } = Attribute;
-const { and } = Predicate;
 
 /**
  * Parse a "span" (colspan/rowspan) attribute on table cell according to specs.
  * @see https://html.spec.whatwg.org/multipage/tables.html#algorithm-for-processing-rows
- * Steps 8 and 9.
+ *      Steps 8 and 9.
  */
 export function parseSpan(
   element: Element,
@@ -28,27 +25,6 @@ export function parseSpan(
     .getOr(Ok.of(failed))
     .getOr(failed);
 }
-
-// Bad copied form alfa-aria accessible name computation (aria-labelledby). Move to DOM helpers?
-export function resolveReferences(
-  node: Node,
-  references: Iterable<string>
-): Array<Element> {
-  const elements: Array<Element> = [];
-
-  for (const id of references) {
-    const element = node
-      .descendants()
-      .find(and(Element.isElement, (element) => element.id.includes(id)));
-
-    if (element.isSome()) {
-      elements.push(element.get());
-    }
-  }
-
-  return elements;
-}
-// End copied from accessible name computation.
 
 /**
  * @see https://html.spec.whatwg.org/multipage/tables.html#empty-cell
